Fix implicit conversion from python lists to numpy vector
only works for double vectors

+++ python/dartpy/gui/osg/ImGuiViewer.cpp
@@ -50,7 +50,7 @@ void ImGuiViewer(py::module& m)
       osg::ref_ptr<dart::gui::osg::ImGuiViewer>>(m, "ImGuiViewer")
       .def(::py::init<>())
       .def(
-          ::py::init([](const Eigen::Vector4f& clearColor) {
+          ::py::init([](const Eigen::Vector4d& clearColor) {
             return new ::dart::gui::osg::ImGuiViewer(eigToOsgVec4f(clearColor));
           }),
           ::py::arg("clearColor"))
